+++ custom_components/prix_carburant/tools.py
@@ -1,169 +1,223 @@
 """Tools for Prix Carburant."""
-import csv
+from asyncio import timeout
+import json
 import logging
-from math import atan2, cos, radians, sin, sqrt
-from typing import Optional
-import urllib.request
-import zipfile
+import os
+from socket import gaierror
 
-import xmltodict
+from aiohttp import ClientError, ClientSession
 
 from homeassistant.const import ATTR_LATITUDE, ATTR_LONGITUDE, ATTR_NAME
-from homeassistant.util.unit_system import METRIC_SYSTEM, UnitSystem
 
 from .const import (
     ATTR_ADDRESS,
+    ATTR_BRAND,
     ATTR_CITY,
-    ATTR_DISTANCE,
     ATTR_FUELS,
     ATTR_POSTAL_CODE,
     ATTR_PRICE,
     ATTR_UPDATED_DATE,
+    FUELS,
 )
 
 _LOGGER = logging.getLogger(__name__)
 
-STATIONS_DATA_URL = "https://static.data.gouv.fr/resources/prix-des-carburants-en-france/20181117-111538/active-stations.csv"
-STATIONS_TARIFS_URL = "https://donnees.roulez-eco.fr/opendata/instantane"
+PRIX_CARBURANT_API_URL = "https://data.economie.gouv.fr/api/explore/v2.1/catalog/datasets/prix-des-carburants-en-france-flux-instantane-v2/records"
+STATIONS_NAME_FILE = "stations_name.json"
 
 
 class PrixCarburantTool:
     """Prix Carburant class with stations information."""
 
     def __init__(
-        self, user_latitude: float = None, user_longitude: float = None
+        self,
+        user_latitude: float,
+        user_longitude: float,
+        user_range: int,
+        time_zone: str = "Europe/Paris",
+        request_timeout: int = 30,
+        session: ClientSession | None = None,
     ) -> None:
         """Init tool."""
         self._user_latitude = user_latitude
         self._user_longitude = user_longitude
-        self._stations_names = {}
-        self._stations_tarifs = {}
+        self._user_range = user_range
+        self._user_time_zone = time_zone
+        self._stations_names: dict[str, dict] = {}
+        self._stations_data: dict[str, dict] = {}
 
-        self._get_stations_names()
+        self._request_timeout = request_timeout
+        self._session = session
+        self._close_session = False
+
+        if self._session is None:
+            self._session = ClientSession()
+            self._close_session = True
 
     @property
     def stations(self) -> dict:
         """Return stations information."""
-        return self._stations_tarifs
+        return self._stations_data
 
-    def update(self) -> None:
-        """Update tarifs."""
-        if not self._stations_names:
-            raise Exception("call load() first")
-        self._get_stations_tarifs()
-        # add station name
-        for station in self._stations_tarifs.items():
-            station_id = station[0]
-            if station_id in self._stations_names:
-                self._stations_tarifs[station_id][ATTR_NAME] = (
-                    str(self._stations_names[station_id])
-                    .title()
-                    .replace("Station ", "")
-                )
-
-    def _get_stations_names(self) -> None:
-        filehandle, _ = urllib.request.urlretrieve(STATIONS_DATA_URL)
-        data = {}
-        with open(filehandle, newline="", encoding="UTF-8") as file:
-            spamreader = csv.reader(file, delimiter=",", quotechar='"')
-            for row in spamreader:
-                data.update({row[0]: row[1]})
-        self._stations_names = data
-
-    def _get_stations_tarifs(self) -> None:
-        """Update data from all stations."""
-        data = {}
-        filehandle, _ = urllib.request.urlretrieve(STATIONS_TARIFS_URL)
-        with zipfile.ZipFile(filehandle, "r") as zip_file_object:
-            with zip_file_object.open(zip_file_object.namelist()[0]) as file:
-                xml_content = file.read()
-                raw_content = xmltodict.parse(xml_content)
-                for station in raw_content["pdv_liste"]["pdv"]:
-                    try:
-                        latitude = float(station["@latitude"]) / 100000
-                        longitude = float(station["@longitude"]) / 100000
-                        distance = self._get_distance_from_user(latitude, longitude)
-                        data.update(
-                            {
-                                station["@id"]: {
-                                    ATTR_LATITUDE: latitude,
-                                    ATTR_LONGITUDE: longitude,
-                                    ATTR_DISTANCE: distance,
-                                    ATTR_ADDRESS: station["adresse"],
-                                    ATTR_POSTAL_CODE: station["@cp"],
-                                    ATTR_CITY: station["ville"],
-                                    ATTR_NAME: "undefined",
-                                    ATTR_FUELS: {},
-                                }
-                            }
-                        )
-                        if "prix" in station:
-                            for fuel in station["prix"]:
-                                fuel_info = (
-                                    fuel if isinstance(fuel, dict) else station["prix"]
-                                )
-                                data[station["@id"]][ATTR_FUELS].update(
-                                    {
-                                        fuel_info["@nom"]: {
-                                            ATTR_UPDATED_DATE: fuel_info["@maj"],
-                                            ATTR_PRICE: fuel_info["@valeur"],
-                                        }
-                                    }
-                                )
-                    except (KeyError, TypeError) as error:
-                        _LOGGER.error(
-                            "Error while getting station %s information: %s",
-                            station["@id"],
-                            error,
-                        )
-
-        self._stations_tarifs = data
-
-    def _get_distance_from_user(
-        self, latitude: float, longitude: float, units: UnitSystem = METRIC_SYSTEM
-    ) -> Optional[float]:
-        """Return the distance with user."""
-        if latitude and longitude and self._user_latitude and self._user_longitude:
-            return _get_distance(
-                self._user_longitude, self._user_latitude, longitude, latitude, units
-            )
-        return None
-
-    def get_near_stations(
+    async def _request_api(
         self,
-        latitude: float,
-        longitude: float,
-        distance: int,
-        units: UnitSystem = METRIC_SYSTEM,
-    ) -> list:
-        """Return list of station near the location."""
-        near_stations_ids = []
-        for station_id, station_info in self._stations_tarifs.items():
-            if station_info[ATTR_LATITUDE] and station_info[ATTR_LONGITUDE]:
-                station_distance = _get_distance(
-                    latitude,
-                    longitude,
-                    float(station_info[ATTR_LATITUDE]),
-                    float(station_info[ATTR_LONGITUDE]),
-                    units,
+        params: dict,
+    ) -> dict:
+        """Make a request to the JSON API."""
+        try:
+            params.update(
+                {
+                    "lang": "fr",
+                    "timezone": self._user_time_zone,
+                }
+            )
+            async with timeout(self._request_timeout):
+                response = await self._session.request(  # type: ignore[union-attr]
+                    method="GET", url=PRIX_CARBURANT_API_URL, params=params
                 )
-                if station_distance <= distance:
-                    near_stations_ids.append(station_id)
-        return near_stations_ids
+                content = await response.json()
+
+                if response.status == 200 and "results" in content:
+                    response.close()
+                    return content
+
+                raise PrixCarburantToolRequestError(
+                    f"API request error {response.status}: {content}"
+                )
+
+        except TimeoutError as exception:
+            raise PrixCarburantToolCannotConnectError(
+                "Timeout occurred while connecting to Prix Carburant API."
+            ) from exception
+        except (ClientError, gaierror) as exception:
+            raise PrixCarburantToolCannotConnectError(
+                "Error occurred while communicating with the Prix Carburant API."
+            ) from exception
+
+    async def init_stations_data(self) -> None:
+        """Get data from all stations."""
+        _LOGGER.debug("load stations names from: %s", STATIONS_NAME_FILE)
+        with open(
+            os.path.join(
+                os.path.dirname(os.path.abspath(__file__)), STATIONS_NAME_FILE
+            ),
+            encoding="UTF-8",
+        ) as file:
+            self._stations_names = json.load(file)
+
+        data = {}
+        _LOGGER.debug("call %s API for station data retrieval", PRIX_CARBURANT_API_URL)
+        response_count = await self._request_api(
+            {
+                "select": "id",
+                "where": f"distance(geom, geom'POINT({self._user_longitude} {self._user_latitude})', {self._user_range}km)",
+                "limit": 1,
+            }
+        )
+        stations_count = response_count["total_count"]
+        _LOGGER.debug("%s stations returned by the API", stations_count)
+
+        for query_offset in range(0, stations_count, 100):
+            query_limit = (
+                100
+                if query_offset < stations_count - 100
+                else stations_count - query_offset
+            )
+            _LOGGER.debug(
+                "query stations from %s to %s/%s",
+                query_offset,
+                query_limit,
+                stations_count,
+            )
+            async with timeout(self._request_timeout):
+                response = await self._request_api(
+                    {
+                        "select": "id,latitude,longitude,cp,adresse,ville",
+                        "where": f"distance(geom, geom'POINT({self._user_longitude} {self._user_latitude})', {self._user_range}km)",
+                        "offset": query_offset,
+                        "limit": query_limit,
+                    }
+                )
+            for station in response["results"]:
+                try:
+                    data.update(
+                        {
+                            station["id"]: {
+                                ATTR_LATITUDE: float(station["latitude"]) / 100000,
+                                ATTR_LONGITUDE: float(station["longitude"]) / 100000,
+                                ATTR_ADDRESS: station["adresse"],
+                                ATTR_POSTAL_CODE: station["cp"],
+                                ATTR_CITY: station["ville"],
+                                ATTR_NAME: "undefined",
+                                ATTR_BRAND: None,
+                                ATTR_FUELS: {},
+                            }
+                        }
+                    )
+                    # add station name if existing in local data
+                    if str(station["id"]) in self._stations_names:
+                        data[station["id"]][ATTR_NAME] = (
+                            str(self._stations_names[str(station["id"])]["Nom"])
+                            .title()
+                            .replace("Station ", "")
+                        )
+                        data[station["id"]][ATTR_BRAND] = str(
+                            self._stations_names[str(station["id"])]["Marque"]
+                        ).title()
+
+                except (KeyError, TypeError) as error:
+                    _LOGGER.error(
+                        "Error while getting station %s information: %s",
+                        station.get("id", "no ID"),
+                        error,
+                    )
+
+        self._stations_data = data
+
+    async def update_stations_prices(self) -> None:
+        """Update tarifs of specified stations."""
+        _LOGGER.debug("call %s API for station data retrieval", PRIX_CARBURANT_API_URL)
+        for station_id, station_data in self._stations_data.items():
+            _LOGGER.debug(
+                "Update fuel prices for station id %s: %s",
+                station_id,
+                station_data[ATTR_NAME],
+            )
+
+            query_select = ",".join(
+                [f"{f.lower()}_prix" for f in FUELS]
+                + [f"{f.lower()}_maj" for f in FUELS]
+            )
+
+            response = await self._request_api(
+                {
+                    "select": query_select,
+                    "where": f"id={station_id}",
+                    "limit": 1,
+                }
+            )
+            if response["total_count"] != 1:
+                _LOGGER.error(
+                    "%s stations returned, must be 1", response["total_count"]
+                )
+                continue
+            new_prices = response["results"][0]
+            for fuel in FUELS:
+                fuel_key = fuel.lower()
+                if new_prices[f"{fuel_key}_prix"]:
+                    station_data[ATTR_FUELS].update(
+                        {
+                            fuel: {
+                                ATTR_UPDATED_DATE: new_prices[f"{fuel_key}_maj"],
+                                ATTR_PRICE: new_prices[f"{fuel_key}_prix"],
+                            }
+                        }
+                    )
 
 
-def _get_distance(
-    lon1: float, lat1: float, lon2: float, lat2: float, units: UnitSystem
-) -> float:
-    """Get distance from 2 locations."""
-    earth_radius = 6371 if units == METRIC_SYSTEM else 3956
+class PrixCarburantToolCannotConnectError(Exception):
+    """Exception to indicate an error in connection."""
 
-    # convert decimal degrees to radians
-    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])
 
-    # haversine formula
-    dlon = lon2 - lon1
-    dlat = lat2 - lat1
-    calcul_a = sin(dlat / 2) ** 2 + cos(lat1) * cos(lat2) * sin(dlon / 2) ** 2
-    calcul_c = 2 * atan2(sqrt(calcul_a), sqrt(1 - calcul_a))
-    return round(calcul_c * earth_radius, 2)
+class PrixCarburantToolRequestError(Exception):
+    """Exception to indicate an error with an API request."""